Now percent-encoding with urllib.quote.
In particular, we use urllib.quote(path_and_querystring, '?=')
because we always have '?' and '=' in the url string functioning
in their reserved-character roles.

Also, prior to this commit we were appending API_BASE_URL in
several different functions.  With this commit we build the
full URL only in _make_request.